refactor: page 위치 변경
